new edit

+++ PPE.API/Business/ViewModels/E6ViewModel.cs
@@ -9,18 +9,18 @@ namespace PPE.API.Business.ViewModels
         public DateTime Date {get; set;}
         public int EtudiantId {get; set;}
         public int JuryId {get; set;}
-        public int C1 {get; set;}
-        public int C2 {get; set;}
-        public int C3 {get; set;}
-        public int C4 {get; set;}
-        public int NotePhase1 {get; set;}
-        public int C5 {get; set;}
-        public int C6 {get; set;}
-        public int C7 {get; set;}
-        public int NotePhase2 {get; set;}
-        public int P2 {get; set;}
-        public int P1 {get; set;}
-        public int Resultat {get; set;}
+        public float C1 {get; set;}
+        public float C2 {get; set;}
+        public float C3 {get; set;}
+        public float C4 {get; set;}
+        public float NotePhase1 {get; set;}
+        public float C5 {get; set;}
+        public float C6 {get; set;}
+        public float C7 {get; set;}
+        public float NotePhase2 {get; set;}
+        public float P2 {get; set;}
+        public float P1 {get; set;}
+        public float Resultat {get; set;}
 
     }
 }
