+++ PPE.API/Migrations/PPEAPIContextModelSnapshot.cs
@@ -122,31 +122,31 @@ namespace PPE.API.Migrations
                         .HasColumnName("ID")
                         .HasColumnType("INTEGER");
 
-                    b.Property<int>("C1")
+                    b.Property<float>("C1")
                         .HasColumnName("C1")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C2")
+                    b.Property<float>("C2")
                         .HasColumnName("C2")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C3")
+                    b.Property<float>("C3")
                         .HasColumnName("C3")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C4")
+                    b.Property<float>("C4")
                         .HasColumnName("C4")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C5")
+                    b.Property<float>("C5")
                         .HasColumnName("C5")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C6")
+                    b.Property<float>("C6")
                         .HasColumnName("C6")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("C7")
+                    b.Property<float>("C7")
                         .HasColumnName("C7")
                         .HasColumnType("REAL");
 
@@ -162,23 +162,23 @@ namespace PPE.API.Migrations
                         .HasColumnName("JuryId")
                         .HasColumnType("INTEGER");
 
-                    b.Property<int>("NotePhase1")
+                    b.Property<float>("NotePhase1")
                         .HasColumnName("NotePhase1")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("NotePhase2")
+                    b.Property<float>("NotePhase2")
                         .HasColumnName("NotePhase2")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("P1")
+                    b.Property<float>("P1")
                         .HasColumnName("NotePhase1")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("P2")
+                    b.Property<float>("P2")
                         .HasColumnName("NotePhase1")
                         .HasColumnType("REAL");
 
-                    b.Property<int>("Resultat")
+                    b.Property<float>("Resultat")
                         .HasColumnName("Resultat")
                         .HasColumnType("REAL");
